Day 3 refactor: Set and Grid
Go does not have a Set in std lib, and although it's quite trivial to
implement, it feels too cumbersome to re-type this... so introducing
basic collections with a Set (union etc... to follow).
Implementing a grid was ok-ish, but still extracted some basic
functionality in a lib.
Time will change these things, I'm sure, but for now, it feels good as a
base for a future puzzle...