Refactoring + optimize day 4
I wanted to keep the commits atomic, as in that every day should still
work in every rollback, force of habit I guess.

I brainfarted quite a lot when solving Day 4, part 2; first try I was
having trouble with recursion (sample data worked, but real input seem
to take ages (stopped run after a minute or so).  I then tried to go
without recursion, but I still brain farted and started to see recursion
everywhere.  In a slight moment of clarity I found a solution, but it
was still based in keeping the winning and played numbers etc... which
resulted in a working but overly complex looping over and over again.
But, writing out that solution made me see the essential parts needed
for solving the puzzle.

I expanded the collections.Set with `Intersection`, which was nice for
day 4.  Then I realized that I could generate the indices of the
produced copies of cards if I had an index of the current card when
reading the file, resulting in a function signature change of
`ProcessLine()` in the `LineProcessor` interface.  I could of course
just use a global int in day 4, but this is nicer, and very much
reusable in the future, so bullet biting it is.