Day 2: Fix overzealous parsing
Consider it a bug, but in the context of the puzzle, it did not matter,
just generated too much arrays; ie. 1 array for each number of cubes
instead of 1 per grab.
For instance, in following line:
Game 1: 3 blue, 4 red; 1 red, 2 green, 6 blue; 2 green
The grabs would originally be parsed as:
[0 0 3] [4 0 0] [1 0 0] [0 2 0] [0 0 6] [0 2 0]
So checking max/min values, was still ok, but the data structure
should've been:
[4 0 3] [1 2 6] [0 2 0]
This change fixes that