Solving day 6 as quadratic function
This runs an order of magnitude faster; quick `time` of 20 runs showed
ca. 10-20x.
Not that it really mattered, because figuring this out took on order of
magnitude longer than just bruteforce-running part2 :-D, but it was a
fun exercise.

The exact calculation of the interval threw me off, since I had not
taken (record + 1) into account and only Floor()'d the interval ends.
Which all worked just fine _except_ for the last race of the sample
data.

Thx to Jelle/Yasmin for reminding me of quadratic functions!