Add basic main.go template
I've been doing a lot of copypasting, so I'm putting up a basic template
for a day's `main.go`.
I think that what I did on day 9, ie. calculating in the `ProcessLine()`
function is preferable since you only iterate over the lines once, most
of the times.
Most of the times, there's also a lot of overlap between the two parts
of day.  I find myself copying a lot of stuff:

  - `part1()` and `part2()`
  - `Solution1{}` and -2
  - Their corresponding `ProcessLine()` functions
  - ...

I think we can go for calculating an answer for the parts separately and
just print those.
This might not always be the case, of course, but we'll see what we'll
do then.  Some basic refactoring between part 1 and 2 will most probably
go a long way.