Revisit Day 1
I liked the approach of part 2 in 01-alt/ but the parsing of part 1
seemed a bit too much.
For part 1, a regex seems the way to go.
I liked the conversion to rune better than doing `atoi()`, but I'm sure
that this is the "best" way to do it.

Also tried to go with the main.go template introduced in [3a71d35],
which seems to work _way_ better.

I'm keeping the original solution as a testimony of how one can be so
lost in the coding woods :-)