Day 10, part 2
Damn, that was hard.
Long story short: I wanted to enter the loop and do a flood fill,
disregarding for a minute that I had to check/lookup that algorithm,
both entering and the touching loop edges were throwing me off.

But drawing out a lot of row/col loops and checking the path to enter
the loop, it seemed the you could count the numbers of times you'd cross
the loop by going to the right edge of the grid.

When coding, I went down a few rabbit holes, but to summarize:

  - I forgot to find the pipe on the starting position, this required me
    to expand my little collections.Set
  - Finding the chars to check so you do not count U and ∩ shapes took
    some trial and error.
  - And, most importantly; I forgot to clean up the grid after I had
    found the loop!!! (this was a serious :facepalm:) I kept hitting way
    to large numbers!
  - `c` for 'column' and `c` for 'char' did not help either.