Day 12, part 1
I took the "TRY ALL THE COMBINATIONS" - brute force route quite early,
but was going in circles on how to count correctly... Finally made it by
introducing the `decrement()` func.
This is naive a.f. really and runs in about 5 seconds, but I really
wanted to finish it this way :-)
Pretty sure part2 is gonna be a nono with this implementation, but we'll
see...