Day 18, part 2
This was a long journey; after learning about Shoelace in part 1, I
learned the adjustments I did for adding the boundary path to the area,
resembled Pick's Theorem... End result being that we use shoelace +
Pick's theorem to have the _complete_ inner area.  Once we have that, we
just add the path length to it (since the length of the path is the
exact area)

Doing part 2, I noticed the time it took to do the actual moves, like in
part1 with the `move()` function, see [3322ab3] was taking quite some
time (ca. 10s).  Noticed that we don't really need to store the whole
path, we just need the length of it.  And since we know the number of
steps, creating the coordinates of the corners and calculating the path
length is quite trivial.

Should've gone to
https://en.wikipedia.org/wiki/Shoelace_formula#Other_formulas earlier
instead of relying on my 1-1 implementation of
https://www.themathdoctors.org/polygon-coordinates-and-areas/
(Although thanks to the latter, I knew about "shoelace" terminology)